Refactor config imports to use submodule entrypoints
